+++ lib/clean_framework.dart
@@ -11,5 +11,6 @@ export 'package:clean_framework/src/logger.dart';
 export 'package:clean_framework/src/open_feature/open_feature.dart';
 export 'package:clean_framework/src/routing/app_router.dart';
 export 'package:clean_framework/src/utilities/deserializer.dart';
+export 'package:clean_framework/src/utilities/file.dart';
 export 'package:clean_framework/src/widgets/widgets.dart';
 export 'package:either_dart/either.dart';