+++ lib/src/open_feature/src/core/enums/flag_value_type.dart
@@ -1,6 +1,14 @@
+/// The value type of a flag.
 enum FlagValueType {
+  /// Boolean Value.
   boolean,
+
+  /// String Value.
   string,
+
+  /// Numeric Value.
   number,
+
+  /// Object Value.
   object,
 }