+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# [1.3.2] - Jun 29, 2022
+# [1.4.0] - Aug 3, 2022
+- Added `push` method to the router.
+- Added `AppRouter.page` and `AppRouter.custom`.
 - Added `Deserializer` to help with deserializing JSON data.
+- Added support for form data and multipart request through `RestExternalInterface`.
 - Updated dependencies to latest version.
 - [**Breaking**] Removed `valueType` from **FeatureBuilder**.